requirements.txt
PyPDF2==3.0.1
urllib3<2.0
langchain>=0.0.310
langchain-community>=0.0.1
langchain-openai>=0.0.1
openai>=0.27.8
gradio==3.40.1
python-dotenv==1.0.0